Move libsndfile sample load code into Buffer

+++ signum/constants.h
@@ -1,1 +1,1 @@
-typedef double sample;
+typedef float sample;
